remove MarkdownCopyAsHtmlFormattedAction fix NPE
