+++ gradle.properties
@@ -3,7 +3,7 @@
 pluginGroup = io.github.md2conf
 pluginName = idea-markdown-copy-as-plugin
 # SemVer format -> https://semver.org
-pluginVersion = 0.0.5
+pluginVersion = 0.0.6
 
 # Supported build number ranges and IntelliJ Platform versions -> https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
 pluginSinceBuild = 213
@@ -11,7 +11,7 @@ pluginUntilBuild =
 
 # IntelliJ Platform Properties -> https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html#intellij-extension
 platformType = IC
-platformVersion = 2022.1.1
+platformVersion = 2022.3
 
 # Plugin Dependencies -> https://plugins.jetbrains.com/docs/intellij/plugin-dependencies.html
 # Example: platformPlugins = com.intellij.java, com.jetbrains.php:203.4449.22
