+++ src/main/resources/META-INF/plugin.xml
@@ -16,13 +16,5 @@
             <add-to-group group-id="EditorPopupMenu" anchor="after" relative-to-action="$Copy"/>
         </action>
 
-        <action id="com.github.qwazer.ideamarkdowncopyas.action.MarkdownCopyAsHtmlFormattedAction"
-                class="com.github.qwazer.ideamarkdowncopyas.action.MarkdownCopyAsHtmlFormattedAction"
-                text="Copy Markdown as HTML Formatted"
-                description="Copy markdown as HTML formatted">
-            <add-to-group group-id="CutCopyPasteGroup" anchor="after" relative-to-action="CopyPaths"/>
-            <add-to-group group-id="EditorPopupMenu" anchor="after" relative-to-action="$Copy"/>
-        </action>
-
     </actions>
 </idea-plugin>
